// src/types/data.type.ts
export type Data = {
  TotalLayer: number
  bedTemp0: string
  curPosition: string
  layer: number
  nozzleTemp: string
  printFileName: string
  printJobTime: number
  printLeftTime: number
  printProgress: number
  printStartTime: number
  state: number
  targetBedTemp0: number
  targetNozzleTemp: number
}