// src/types/ws-message.type.ts
type WsMessageErr = {
  errcode: number
  key: number
}

export type WsMessage = {
  TotalLayer?: number
  accelToDecelLimits?: number
  accelerationLimits?: number
  aiDetection?: number
  aiFirstFloor?: number
  aiPausePrint?: number
  aiSw?: number
  autoLevelResult?: string
  autohome?: string
  auxiliaryFanPct?: number
  bedTemp0?: string
  bedTemp1?: string
  bedTemp2?: string
  bedTempAutoPid?: number
  boxTemp?: number
  caseFanPct?: number
  connect?: number
  cornerVelocityLimits?: number
  curFeedratePct?: number
  curFlowratePct?: number
  curPosition?: string
  dProgress?: number
  deviceState?: number
  enableSelfTest?: number
  err?: WsMessageErr
  fan?: number
  fanAuxiliary?: number
  fanCase?: number
  hostname?: string
  layer?: number
  lightSw?: number
  materialDetect?: number
  materialStatus?: number
  maxBedTemp?: number
  maxNozzleTemp?: number
  model?: string
  modelFanPct?: number
  modelVersion?: string
  nozzleMoveSnapshot?: number
  nozzleTemp?: string
  nozzleTempAutoPid?: number
  powerLoss?: number
  pressureAdvance?: string
  printFileName?: string
  printId?: string
  printJobTime?: number
  printLeftTime?: number
  printProgress?: number
  printStartTime?: number
  realTimeFlow?: string
  realTimeSpeed?: string
  repoPlrStatus?: number
  smoothTime?: string
  state?: number
  targetBedTemp0?: number
  targetBedTemp1?: number
  targetBedTemp2?: number
  targetNozzleTemp?: number
  tfCard?: number
  upgradeStatus?: number
  usedMaterialLength?: number
  velocityLimits?: number
  video?: number
  video1?: number
  videoElapse?: number
  videoElapseFrame?: number
  videoElapseInterval?: number
  withSelfTest?: number
}